build: Add GObject introspection support
This bumps our autoconf requirement to 2.62 (from 2.59c) and disables
automake portability warnings.

It does not introduce a hard dependency on gobject-introspection; but if
available, it requires version 1.30.0. Support can be enabled with
--enable-introspection, and is required to make distcheck.

The generated introspection file has not been checked thoroughly, and
may expose internal symbols. A lot of new annotations are needed to make
many of the symbols usable.