Fix T121 Libnice not picking up UDP packets
Set winsock options to get around a [known bug in windows][bug].
Without this patch, UDP sockets may falsely return a `WSAECONNRESET`
when trying to read data. Once libnice receives a `WSAECONNRESET` on
a socket, it detaches the source and cleans up, as it should, so these
false positives (the socket is still there and can receive data) cause
libnice to ditch perfectly good connections.

[T121 - Bug report][report]

[bug]: https://support.microsoft.com/en-us/kb/263823
[report]: http://phabricator.freedesktop.org/T121