conncheck.c: Fix negotiation problem
When a peer reflexive candidate is discovered, it is added to the
list of connection checks, but is is not nominated by default.

This causes a problem if parent pair is already nominated and
the remote peer does not send more requests because it has already
selected a candidate. (i.e.: This is the case of WebRtc in Firefox)